Move zocalo gridscan integration to preprocessor
