+++ src/mx_bluesky/common/experiment_plans/common_flyscan_xray_centre_plan.py
@@ -1,28 +1,20 @@
 from __future__ import annotations
 
 import dataclasses
-from collections.abc import Callable, Sequence
+from collections.abc import Callable
 from functools import partial
 
 import bluesky.plan_stubs as bps
 import bluesky.preprocessors as bpp
-import numpy as np
 from bluesky.protocols import Readable
 from bluesky.utils import FailedStatus, MsgGenerator
-from dodal.common.beamlines.commissioning_mode import read_commissioning_mode
 from dodal.devices.fast_grid_scan import (
     FastGridScanCommon,
     FastGridScanThreeD,
     GridScanInvalidError,
 )
-from dodal.devices.zocalo import ZocaloResults
-from dodal.devices.zocalo.zocalo_results import (
-    XrcResult,
-    get_full_processing_results,
-)
 
 from mx_bluesky.common.experiment_plans.inner_plans.do_fgs import (
-    ZOCALO_STAGE_GROUP,
     kickoff_and_complete_gridscan,
 )
 from mx_bluesky.common.experiment_plans.inner_plans.read_hardware import (
@@ -30,19 +22,16 @@ from mx_bluesky.common.experiment_plans.inner_plans.read_hardware import (
 )
 from mx_bluesky.common.parameters.constants import (
     DocDescriptorNames,
-    GridscanParamConstants,
     PlanGroupCheckpointConstants,
     PlanNameConstants,
 )
 from mx_bluesky.common.parameters.device_composites import FlyScanEssentialDevices
 from mx_bluesky.common.parameters.gridscan import SpecifiedThreeDGridScan
 from mx_bluesky.common.utils.exceptions import (
-    CrystalNotFoundError,
     SampleError,
 )
 from mx_bluesky.common.utils.log import LOGGER
 from mx_bluesky.common.utils.tracing import TRACER
-from mx_bluesky.common.xrc_result import XRayCentreResult
 
 
 @dataclasses.dataclass
@@ -55,16 +44,12 @@ class BeamlineSpecificFGSFeatures:
         ..., MsgGenerator
     ]  # Eventually replace with https://github.com/DiamondLightSource/mx-bluesky/issues/819
     read_during_collection_plan: Callable[..., MsgGenerator]
-    get_xrc_results_from_zocalo: bool
 
 
 def generic_tidy(xrc_composite: FlyScanEssentialDevices, wait=True) -> MsgGenerator:
-    """Tidy Zocalo and turn off Eiger dev/shm. Ran after the beamline-specific tidy plan"""
+    """Turn off Eiger dev/shm. Ran after the beamline-specific tidy plan"""
 
-    LOGGER.info("Tidying up Zocalo")
     group = "generic_tidy"
-    # make sure we don't consume any other results
-    yield from bps.unstage(xrc_composite.zocalo, group=group)
 
     # Turn off dev/shm streaming to avoid filling disk, see https://github.com/DiamondLightSource/hyperion/issues/1395
     LOGGER.info("Turning off Eiger dev/shm streaming")
@@ -84,7 +69,6 @@ def construct_beamline_specific_fast_gridscan_features(
     fgs_motors: FastGridScanCommon,
     signals_to_read_pre_flyscan: list[Readable],
     signals_to_read_during_collection: list[Readable],
-    get_xrc_results_from_zocalo: bool = False,
 ) -> BeamlineSpecificFGSFeatures:
     """Construct the class needed to do beamline-specific parts of the XRC FGS
 
@@ -93,7 +77,7 @@ def construct_beamline_specific_fast_gridscan_features(
         Ran directly before kicking off the gridscan.
 
         tidy_plan (Callable): Tidy up states of devices. Ran at the end of the flyscan, regardless of
-        whether or not it finished successfully. Zocalo and Eiger are cleaned up separately
+        whether or not it finished successfully. Eiger is cleaned up separately
 
         set_flyscan_params_plan (Callable): Set PV's for the relevant Fast Grid Scan dodal device
 
@@ -104,9 +88,6 @@ def construct_beamline_specific_fast_gridscan_features(
 
         signals_to_read_during_collection (Callable): Signals which will be read and saved as a bluesky event
         document whilst the gridscan motion is in progress
-
-        get_xrc_results_from_zocalo (bool): If true, fetch grid scan results from zocalo after completion, as well as
-        update the ispyb comment field with information about the results. See _fetch_xrc_results_from_zocalo
     """
     read_pre_flyscan_plan = partial(
         read_hardware_plan,
@@ -127,7 +108,6 @@ def construct_beamline_specific_fast_gridscan_features(
         fgs_motors,
         read_pre_flyscan_plan,
         read_during_collection_plan,
-        get_xrc_results_from_zocalo,
     )
 
 
@@ -148,13 +128,14 @@ def common_flyscan_xray_centre(
         centering results.
 
     With a minimum set of devices and parameters, prepares for; performs; and tidies up a flyscan
-    x-ray-center plan. This includes: Configuring desired triggering; writing nexus files; triggering zocalo;
+    x-ray-center plan. This includes: Configuring desired triggering; writing nexus files;
     reading hardware before and during the scan; and tidying up devices after
-    the plan is complete. Optionally fetch results from zocalo after completing the grid scan.
+    the plan is complete. Results can be fetched via zocalo by using the use_gridscan_with_zocalo_decorator
+    preprocessor.
 
     This plan will also push data to ispyb when used with the ispyb_activation_decorator.
 
-    There are a few other useful decorators to use with this plan, see: verify_undulator_gap_before_run_decorator, common/preprocessors/preprocessors.py
+    There are a few other useful decorators to use with this plan, see: verify_undulator_gap_before_run_decorator, transmission_and_xbpm_feedback_for_collection_decorator
     """
 
     def _overall_tidy():
@@ -181,79 +162,14 @@ def common_flyscan_xray_centre(
             yield from beamline_specific.setup_trigger_plan(fgs_composite, parameters)
 
             LOGGER.info("Starting grid scan")
-            yield from bps.stage(
-                fgs_composite.zocalo, group=ZOCALO_STAGE_GROUP
-            )  # connect to zocalo and make sure the queue is clear
             yield from run_gridscan(fgs_composite, params, beamline_specific)
 
             LOGGER.info("Grid scan finished")
-
-            if beamline_specific.get_xrc_results_from_zocalo:
-                yield from _fetch_xrc_results_from_zocalo(composite.zocalo, parameters)
 
         yield from run_gridscan_and_tidy(composite, parameters, beamline_specific)
 
     composite.eiger.set_detector_parameters(parameters.detector_params)
     yield from _decorated_flyscan()
-
-
-def _fetch_xrc_results_from_zocalo(
-    zocalo_results: ZocaloResults,
-    parameters: SpecifiedThreeDGridScan,
-) -> MsgGenerator:
-    """
-    Get XRC results from the ZocaloResults device which was staged during a grid scan,
-    and store them in XRayCentreEventHandler.xray_centre_results by firing an event.
-
-    The RunEngine must be subscribed to XRayCentreEventHandler for this plan to work.
-    """
-
-    LOGGER.info("Getting X-ray center Zocalo results...")
-
-    yield from bps.trigger(zocalo_results)
-    LOGGER.info("Zocalo triggered and read, interpreting results.")
-    xrc_results = yield from get_full_processing_results(zocalo_results)
-    LOGGER.info(f"Got xray centres, top 5: {xrc_results[:5]}")
-    filtered_results = [
-        result
-        for result in xrc_results
-        if result["total_count"]
-        >= GridscanParamConstants.ZOCALO_MIN_TOTAL_COUNT_THRESHOLD
-    ]
-    discarded_count = len(xrc_results) - len(filtered_results)
-    if discarded_count > 0:
-        LOGGER.info(f"Removed {discarded_count} results because below threshold")
-    if filtered_results:
-        flyscan_results = [
-            _xrc_result_in_boxes_to_result_in_mm(xr, parameters)
-            for xr in filtered_results
-        ]
-    else:
-        commissioning_mode = yield from read_commissioning_mode()
-        if commissioning_mode:
-            LOGGER.info("Commissioning mode enabled, returning dummy result")
-            flyscan_results = [_generate_dummy_xrc_result(parameters)]
-        else:
-            LOGGER.warning("No X-ray centre received")
-            raise CrystalNotFoundError()
-    yield from _fire_xray_centre_result_event(flyscan_results)
-
-
-def _generate_dummy_xrc_result(params: SpecifiedThreeDGridScan) -> XRayCentreResult:
-    com = [params.x_steps / 2, params.y_steps / 2, params.z_steps / 2]
-    max_voxel = [round(p) for p in com]
-    return _xrc_result_in_boxes_to_result_in_mm(
-        XrcResult(
-            centre_of_mass=com,
-            max_voxel=max_voxel,
-            bounding_box=[max_voxel, [p + 1 for p in max_voxel]],
-            n_voxels=1,
-            max_count=10000,
-            total_count=100000,
-            sample_id=params.sample_id,
-        ),
-        params,
-    )
 
 
 @bpp.set_run_key_decorator(PlanNameConstants.GRIDSCAN_MAIN)
@@ -298,49 +214,3 @@ def run_gridscan(
     # in a GDA-happy state.
     if isinstance(beamline_specific.fgs_motors, FastGridScanThreeD):
         yield from bps.abs_set(beamline_specific.fgs_motors.z_steps, 0, wait=False)
-
-
-def _xrc_result_in_boxes_to_result_in_mm(
-    xrc_result: XrcResult, parameters: SpecifiedThreeDGridScan
-) -> XRayCentreResult:
-    fgs_params = parameters.fast_gridscan_params
-    xray_centre = fgs_params.grid_position_to_motor_position(
-        np.array(xrc_result["centre_of_mass"])
-    )
-    # A correction is applied to the bounding box to map discrete grid coordinates to
-    # the corners of the box in motor-space; we do not apply this correction
-    # to the xray-centre as it is already in continuous space and the conversion has
-    # been performed already
-    # In other words, xrc_result["bounding_box"] contains the position of the box centre,
-    # so we subtract half a box to get the corner of the box
-    return XRayCentreResult(
-        centre_of_mass_mm=xray_centre,
-        bounding_box_mm=(
-            fgs_params.grid_position_to_motor_position(
-                np.array(xrc_result["bounding_box"][0]) - 0.5
-            ),
-            fgs_params.grid_position_to_motor_position(
-                np.array(xrc_result["bounding_box"][1]) - 0.5
-            ),
-        ),
-        max_count=xrc_result["max_count"],
-        total_count=xrc_result["total_count"],
-        sample_id=xrc_result["sample_id"],
-    )
-
-
-def _fire_xray_centre_result_event(results: Sequence[XRayCentreResult]):
-    def empty_plan():
-        return iter([])
-
-    yield from bpp.set_run_key_wrapper(
-        bpp.run_wrapper(
-            empty_plan(),
-            md={
-                PlanNameConstants.FLYSCAN_RESULTS: [
-                    dataclasses.asdict(r) for r in results
-                ]
-            },
-        ),
-        PlanNameConstants.FLYSCAN_RESULTS,
-    )
