+++ src/mx_bluesky/common/parameters/device_composites.py
@@ -23,14 +23,12 @@ from dodal.devices.undulator import UndulatorInKeV
 from dodal.devices.xbpm_feedback import XBPMFeedback
 from dodal.devices.zebra.zebra import Zebra
 from dodal.devices.zebra.zebra_controlled_shutter import ZebraShutter
-from dodal.devices.zocalo import ZocaloResults
 
 
 @pydantic.dataclasses.dataclass(config={"arbitrary_types_allowed": True})
 class FlyScanEssentialDevices:
     eiger: EigerDetector
     synchrotron: Synchrotron
-    zocalo: ZocaloResults
     smargon: Smargon
