+++ src/mx_bluesky/hyperion/experiment_plans/hyperion_flyscan_xray_centre_plan.py
@@ -101,7 +101,6 @@ def construct_hyperion_specific_features(
         fgs_motors,
         signals_to_read_pre_flyscan,
         signals_to_read_during_collection,
-        get_xrc_results_from_zocalo=True,
     )
